+++ core/base/src/test/java/org/eclipse/dataspaceconnector/core/base/policy/PolicyEngineImplTest.java
@@ -21,6 +21,7 @@ import org.eclipse.dataspaceconnector.policy.model.Permission;
 import org.eclipse.dataspaceconnector.policy.model.Policy;
 import org.eclipse.dataspaceconnector.policy.model.Prohibition;
 import org.eclipse.dataspaceconnector.spi.agent.ParticipantAgent;
+import org.eclipse.dataspaceconnector.spi.policy.RuleBindingRegistry;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -32,6 +33,7 @@ import static org.eclipse.dataspaceconnector.spi.policy.PolicyEngine.ALL_SCOPES;
 
 class PolicyEngineImplTest {
     private static final String TEST_SCOPE = "test";
+    private RuleBindingRegistry bindingRegistry;
     private PolicyEngineImpl policyEngine;
 
     @Test
@@ -48,6 +50,8 @@ class PolicyEngineImplTest {
 
     @Test
     void validateUnsatisfiedDuty() {
+        bindingRegistry.bind("foo", ALL_SCOPES);
+
         policyEngine.registerFunction(ALL_SCOPES, Duty.class, "foo", (op, rv, duty, context) -> false);
         var agent = new ParticipantAgent(emptyMap(), emptyMap());
 
@@ -62,7 +66,29 @@ class PolicyEngineImplTest {
     }
 
     @Test
+    void validateRuleOutOfScope() {
+        // Verifies that a rule will be filtered if its action is not registered. The constraint is registered but should be filtered since it is contained in the permission.
+        // If the permission is not properly filtered, the constraint will not be fulfilled and raise an exception.
+        bindingRegistry.bind("foo", ALL_SCOPES);
+
+        var agent = new ParticipantAgent(emptyMap(), emptyMap());
+
+        var left = new LiteralExpression("foo");
+        var right = new LiteralExpression("bar");
+        var constraint = AtomicConstraint.Builder.newInstance().leftExpression(left).operator(EQ).rightExpression(right).build();
+
+        var action = Action.Builder.newInstance().type("OUT_OF_SCOPE").build();
+        var permission = Permission.Builder.newInstance().action(action).constraint(constraint).build();
+        var policy = Policy.Builder.newInstance().permission(permission).build();
+
+        // the permission containing the unfulfilled constraint should be filtered, resulting in the policy evaluation succeeding
+        assertThat(policyEngine.evaluate(TEST_SCOPE, policy, agent).succeeded()).isTrue();
+    }
+
+    @Test
     void validateUngrantedPermission() {
+        bindingRegistry.bind("foo", ALL_SCOPES);
+
         policyEngine.registerFunction(ALL_SCOPES, Permission.class, "foo", (op, rv, duty, context) -> false);
         var agent = new ParticipantAgent(emptyMap(), emptyMap());
 
@@ -80,6 +106,8 @@ class PolicyEngineImplTest {
 
     @Test
     void validateTriggeredProhibition() {
+        bindingRegistry.bind("foo", ALL_SCOPES);
+
         policyEngine.registerFunction(ALL_SCOPES, Prohibition.class, "foo", (op, rv, duty, context) -> true);
         var agent = new ParticipantAgent(emptyMap(), emptyMap());
 
@@ -93,6 +121,8 @@ class PolicyEngineImplTest {
 
     @Test
     void validateConstraintFunctionOutOfScope() {
+        bindingRegistry.bind("foo", ALL_SCOPES);
+
         policyEngine.registerFunction("foo", Prohibition.class, "foo", (op, rv, duty, context) -> Assertions.fail("Foo prohibition should be out of scope"));
         policyEngine.registerFunction("bar", Prohibition.class, "foo", (op, rv, duty, context) -> true);
         var agent = new ParticipantAgent(emptyMap(), emptyMap());
@@ -107,6 +137,8 @@ class PolicyEngineImplTest {
 
     @Test
     void validateChildScopeNotVisible() {
+        bindingRegistry.bind("foo", ALL_SCOPES);
+
         policyEngine.registerFunction("bar", Prohibition.class, "foo", (op, rv, duty, context) -> true);
         policyEngine.registerFunction("bar.child", Prohibition.class, "foo", (op, rv, duty, context) -> Assertions.fail("Child prohibition should be out of scope"));
         var agent = new ParticipantAgent(emptyMap(), emptyMap());
@@ -121,6 +153,8 @@ class PolicyEngineImplTest {
 
     @Test
     void validateScopeIsInheritedByChildren() {
+        bindingRegistry.bind("foo", ALL_SCOPES);
+
         policyEngine.registerFunction("bar", Prohibition.class, "foo", (op, rv, duty, context) -> true);
         var agent = new ParticipantAgent(emptyMap(), emptyMap());
 
@@ -134,7 +168,9 @@ class PolicyEngineImplTest {
 
     @Test
     void validateRuleFunctionOutOfScope() {
-        var  action = Action.Builder.newInstance().type("USE").build();
+        bindingRegistry.bind("foo", ALL_SCOPES);
+
+        var action = Action.Builder.newInstance().type("USE").build();
 
         var permission = Permission.Builder.newInstance().action(action).build();
 
@@ -149,6 +185,8 @@ class PolicyEngineImplTest {
 
     @Test
     void validatePreValidator() {
+        bindingRegistry.bind("foo", ALL_SCOPES);
+
         policyEngine.registerPreValidator(ALL_SCOPES, (policy, context) -> false);
         var policy = Policy.Builder.newInstance().build();
         var agent = new ParticipantAgent(emptyMap(), emptyMap());
@@ -160,6 +198,8 @@ class PolicyEngineImplTest {
 
     @Test
     void validatePostValidator() {
+        bindingRegistry.bind("foo", ALL_SCOPES);
+
         policyEngine.registerPostValidator(ALL_SCOPES, (policy, context) -> false);
         var policy = Policy.Builder.newInstance().build();
         var agent = new ParticipantAgent(emptyMap(), emptyMap());
@@ -171,7 +211,8 @@ class PolicyEngineImplTest {
 
     @BeforeEach
     void setUp() {
-        policyEngine = new PolicyEngineImpl();
+        bindingRegistry = new RuleBindingRegistryImpl();
+        policyEngine = new PolicyEngineImpl(new ScopeFilter(bindingRegistry));
     }
 
     private Policy createTestPolicy() {