+++ data-protocols/ids/ids-api-multipart-endpoint-v1/src/main/java/org/eclipse/dataspaceconnector/ids/api/multipart/handler/description/DescriptionRequestHandler.java
@@ -27,6 +27,6 @@ public interface DescriptionRequestHandler {
     @Nullable
     MultipartResponse handle(
             @NotNull DescriptionRequestMessage descriptionRequestMessage,
-            @NotNull Result<ClaimToken> verificationResult,
+            @NotNull ClaimToken claimToken,
             @Nullable String payload);
 }