+++ extensions/api/data-management/asset/src/test/java/org/eclipse/dataspaceconnector/api/datamanagement/asset/transform/AssetDtoToAssetTransformerTest.java
@@ -1,0 +1,53 @@
+/*
+ *  Copyright (c) 2022 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
+ *
+ */
+
+package org.eclipse.dataspaceconnector.api.datamanagement.asset.transform;
+
+import org.eclipse.dataspaceconnector.api.datamanagement.asset.model.AssetDto;
+import org.eclipse.dataspaceconnector.spi.transformer.TransformerContext;
+import org.eclipse.dataspaceconnector.spi.types.domain.asset.Asset;
+import org.junit.jupiter.api.Test;
+
+import java.util.Map;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.mock;
+
+class AssetDtoToAssetTransformerTest {
+
+    private final AssetDtoToAssetTransformer transformer = new AssetDtoToAssetTransformer();
+
+    @Test
+    void inputOutputType() {
+        assertThat(transformer.getInputType()).isNotNull();
+        assertThat(transformer.getOutputType()).isNotNull();
+    }
+
+    @Test
+    void canHandle() {
+        assertThat(transformer.canHandle(AssetDto.Builder.newInstance().build(), Asset.class)).isTrue();
+        assertThat(transformer.canHandle("any other", Asset.class)).isFalse();
+        assertThat(transformer.canHandle(AssetDto.Builder.newInstance().build(), Object.class)).isFalse();
+    }
+
+    @Test
+    void transform() {
+        var context = mock(TransformerContext.class);
+        var assetDto = AssetDto.Builder.newInstance().properties(Map.of("key", "value")).build();
+
+        var asset = transformer.transform(assetDto, context);
+
+        assertThat(asset.getProperties()).containsExactlyEntriesOf(assetDto.getProperties());
+    }
+}